dashboard and comment fix

+++ back-end/Comp1640_Final/DTO/Response/DashboardResponse.cs
@@ -1,9 +1,5 @@
 ﻿namespace Comp1640_Final.DTO.Response
 {
-    //public class DashboardResponse
-    //{
-    //    public Dictionary<int, Dictionary<string, int>> DashboardByYear { get; set; }
-    //}
     public class DashboardResponse
     {
         public string Year { get; set; }
@@ -12,6 +8,6 @@
     public class Values
     {
         public string Faculty { get; set; }
-        public int value { get; set; }
+        public double value { get; set; }
     }
 }
